+++ CHANGELOG.md
@@ -78,6 +78,8 @@ Removed the `Name` paramater aliases. Update usages of the `Name` parameter on t
 * Get-CIisApplication
 * Install-CIisApplication
 
+`Get-CIisAppPool` now writes an error when passed a name and an application pool with that name does not exist. Update
+usages with `-ErrorAction Ignore` to preserve previous behavior.
 
 ## Added
 
@@ -94,6 +96,7 @@ Pass the application name to the `ApplicationPath` parameter.
 configuration (i.e. there's a `<location>` element for it in the applicationHost.config).
 * Function `Remove-CIisConfigurationLocation` for removing a website's or website/virtual path's custom configuration
 (i.e. removes its `<location>` element from applicationHost.config).
+* Many functions now write messages to PowerShell's information stream when they make configuration changes.
 
 ## Changes
 
@@ -101,6 +104,7 @@ configuration (i.e. there's a `<location>` element for it in the applicationHost
 * The `Join-CIisVirtualPath` function's `ChildPath` parameter is now required. Usages that don't have a `ChildPath`
 argument should switch to `ConvertTo-CIisVirtualPath`.
 * Renamed `VirtualPath` parameter on `Get-CIisApplication` and `Install-CIisApplication` to `Path`.
+* `Get-CIisAppPool` now writes an error when passed a name and an application pool with that name does not exist.
 
 ## Fixed